Fixed `ioctl failed: Device or resource busy`
On very rare and very specific occasions `__vpn_connect` may get stuck after a termination request.
However, it's generally safe to just kill it at that stage.
Therefore, that's precisely what we're doing: wait for about 2.5 seconds for it to gracefully finish, and then terminate it if it remains stuck.